backend/requirements.txt
# FastAPI Core Dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database & ORM
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
asyncpg==0.29.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1

# Environment & Configuration
python-dotenv==1.0.0
pydantic==2.5.2
pydantic-settings==2.1.0

# Validation & Data Processing
email-validator==2.1.0

# AI & Machine Learning Dependencies (with compatible versions)
transformers==4.35.2
torch==2.1.1
sentence-transformers==2.2.2
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.2

# NLP & Text Processing
nltk==3.8.1
textblob==0.17.1

# Data Storage & Caching
faiss-cpu==1.7.4
redis==5.0.1

# HTTP & API Client
httpx==0.25.2
requests==2.31.0

# Background Tasks
celery==5.3.4

# Development & Debugging
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0

# Production
gunicorn==21.2.0

# Optional AI providers (can work without API keys)
openai==1.3.8
google-generativeai==0.3.2

# Additional AI libraries
datasets==2.14.6
tokenizers==0.15.0
accelerate==0.24.1
huggingface_hub==0.20.1

# Utilities
aiofiles==23.2.1
joblib==1.3.2